fix: extract restoreScrollbackBuffers to fix max-lines lint error
